fix the rare -xbox/2 case in Spatial and Spatial2
When x is near -xbox/2, its "ix" value may be 2**(px-1) due to
the fact that fx is rounded to +0.5 because of the limited
precision.

For example, fx and ix should have been 0.49999 and 31, but the
program gave 0.5 and 32. This situation has been taken care of
in the past by using double precision in some places, but is
still possible to happen.

This patch doesn't solve the precision issue. What it does is to
"wrap" the ix value in the [0,32) range. fx is still 0.5 but ix
will become 0, instead of the correct 31 value. As a result, this
atom is assigned to its neighboring box. With a resonable cutoff,
and usually a list buffer, I can almost guarantee the constructed
NBList will not be affected. If so, it will only be better than
what we have implemented.